Remake kernel to not spawn redundant blocks/threads
The row was calculated as overall thread number * M.
But since we're running with at least 32 threads per block, and at
least blocks equal to the number of streaming multiprocessors (although
ideally more to take advantage of hardware parallelization/saturating
registers, shared memory, etc.), that means the last thread is at row
(32 * 39 (#SM)) = 1248; this thread will do nothing.
In fact, no thread after 200 / 32 = 6.25 will do anything at all!

The new kernel, while slightly expensive in terms of modulo operations
and lots of `if`s, doesn't do redundant work, and cuts its runtime by a
factor of 10 to be competitive with the OpenMP C++ version, and crush it
when the arrays (and number of iterations) get larger!