Add scipy, cupy, and torch python implementations
My naive python implementation never stood a chance, and neither did
naively iterating through the array with numpy.
My roommate pointed out that the central finite difference scheme used
is essentially just convolving the 2D array by a 3x3 kernel.
So, an easy-ish way to speed up the python implementations is to call
libraries that have already optimized this kind of stuff in a C/C++
backend; Scipy runs on the CPU and should give a good point of reference
for naive C++ implementation, and CuPy and Torch run on the GPU, giving
a point of reference for the CUDA implementation.